Move/Rename whitelister.Whitelist to rich_text.cleaner.HTMLCleaner
